requirements.txt
discord.py=2.3.2
discord.py[voice]=2.3.2
python-dotenv
